authorization and authentication is done  with api & it is implemented eventually

+++ controllers/api/v1/posts_api.js
@@ -19,11 +19,18 @@ module.exports.index = async function(req,res){
 module.exports.destroy = async function(req,res){
     try {
         let post = await Post.findById(req.params.id);
-           post.remove();   
-           await Comment.deleteMany({post: req.params.id});
-           return res.json(200,{
-               message: "Posts and associated comments deleted successfully"
-           });
+            if(post.user == req.user.id)
+            { 
+                post.remove();   
+                await Comment.deleteMany({post: req.params.id});
+                return res.json(200,{
+                    message: "Posts and associated comments deleted successfully"
+                });
+            } else {
+                return res.json(401, {
+                    message: "You can not delete this post!"
+                });
+            }
     
     } catch (err) { 
         console.log(err);
